requirements.txt
Flask==2.2.5
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.0
spacy==3.6.0
transformers==4.35.0
torch==2.2.0
gunicorn==20.1.0
